Change EFI_TPL from size_t to specific uint64_t
Our lord and saviour has asked us the width of our previously somewhat
diffusly applied define, it was typed for size_t, however as our lord
sayeth, we cannot garantee this width, maketh explicit. So we made it 64
bit wide and, so it shall be. EFI_TPL is forever more defined as an
unsigned integer by which the width shall be considered 64 bit wide.

We shall make it so